♻️ #34 refactor: 에디터와 폼 분리
