+++ app/(without-layout)/new/components/Form.tsx
@@ -1,0 +1,61 @@
+"use client";
+
+import { useState } from "react";
+import * as stylex from "@stylexjs/stylex";
+import { useActionState } from "react";
+import Form from "next/form";
+import { publish } from "../actions";
+import { PublishState } from "../types";
+import EditorWrapper from "./editor/Tiptap";
+
+export function PublishForm() {
+  const [state, formAction, pending] = useActionState<PublishState, FormData>(
+    publish,
+    { message: "", title: "", author: "", publisher: "", content: "" }
+  );
+
+  const [content, setContent] = useState("");
+
+  return (
+    <Form action={formAction} {...stylex.props(styles.container)}>
+      <div>
+        <label htmlFor="title">책 제목</label>
+        <input id="title" name="title" required defaultValue={state?.title} />
+      </div>
+      <div>
+        <label htmlFor="author">작가 이름</label>
+        <input id="author" name="author" defaultValue={state?.author} />
+      </div>
+      <div>
+        <label htmlFor="publisher">출판사</label>
+        <input
+          id="publisher"
+          name="publisher"
+          defaultValue={state?.publisher}
+        />
+      </div>
+
+      <EditorWrapper value={content} onChange={setContent} />
+      <input type="hidden" name="content" value={content} />
+
+      <button type="submit" disabled={pending}>
+        발행하기
+      </button>
+      <p aria-live="polite" role="status">
+        {state?.message}
+      </p>
+    </Form>
+  );
+}
+
+const styles = stylex.create({
+  container: {
+    maxWidth: "70vw",
+    margin: "24px auto",
+    border: "1px solid #ccc",
+    padding: "24px",
+    borderRadius: "12px",
+    display: "grid",
+    gap: "12px",
+  },
+});
